:sparkles: feat(frontend): useUploadImage 커스텀훅 생성 (#102)
